requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.4

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth
# bcrypt 4.0.1 is pinned: 4.1+ changed the low-level API in ways that break
# passlib's CryptContext wrapper, causing ValueError on hash/verify calls.
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# ML
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2

# Gemini AI
google-generativeai==0.5.4

# HTTP client (simulator + tests)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
anyio==4.3.0
# aiosqlite provides async SQLite for in-memory test DB (no real Postgres needed)
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
